Fix linking

+++ engine/point.cpp
@@ -2,9 +2,9 @@
 #include <sstream>
 #include <string>
 
-Point::Point(): x(0), y(0), z(0) {}
+Point::Point(): x(0.0f), y(0.0f), z(0.0f) {}
 
-Point::Point(double x, double y, double z) : x(x), y(y), z(z) {}
+Point::Point(float x, float y, float z) : x(x), y(y), z(z) {}
 
 std::string Point::toString() {
     std::ostringstream os;
